pyproject.toml
[project]
name = "sparktest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "delta-spark==3.0.0",
    "pyspark>=3.5.4",
    "requests>=2.32.3",
]
